Refactor for hooks all tests pass
